Add stream join optimization.
If one stream is done, we don't need to store its records anymore, as we will only be matching to the other tree. This is useful for assymetric joins where one table is very small and the other one is very big. Until now, we stored both for the whole time of the join. Now we only store the smaller one, resulting in absurd memory improvement in some cases, as well as much faster processing.